Move theming over to new coreui module
