+++ appNekomp/src/commonMain/kotlin/com/chesire/nekomp/App.kt
@@ -1,5 +1,6 @@
 package com.chesire.nekomp
 
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -9,6 +10,7 @@ import androidx.compose.material3.adaptive.navigationsuite.NavigationSuiteScaffo
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination.Companion.hierarchy
@@ -21,6 +23,7 @@ import androidx.navigation.compose.rememberNavController
 import co.touchlab.kermit.Logger
 import com.chesire.nekomp.core.preferences.ApplicationSettings
 import com.chesire.nekomp.core.preferences.Theme
+import com.chesire.nekomp.core.ui.NekompTheme
 import com.chesire.nekomp.feature.discover.ui.DiscoverScreen
 import com.chesire.nekomp.feature.home.ui.HomeScreen
 import com.chesire.nekomp.feature.library.ui.LibraryScreen
@@ -40,9 +43,19 @@ import org.koin.compose.koinInject
 fun App() {
     Logger.setTag("Nekomp")
 
-    val theme by getKoin().get<ApplicationSettings>().theme.collectAsState(Theme.System)
+    val koin = getKoin()
+    val appSettings = remember { koin.get<ApplicationSettings>() }
+    val theme by appSettings.theme.collectAsState(Theme.System)
 
-    NekompTheme(theme = theme) {
+    val useDarkTheme = when {
+        theme == Theme.System && isSystemInDarkTheme() -> true
+        theme == Theme.System && !isSystemInDarkTheme() -> false
+        theme == Theme.Light -> false
+        theme == Theme.Dark -> true
+        else -> false
+    }
+
+    NekompTheme(useDarkTheme) {
         val appNavController = rememberNavController()
         val isLoggedIn = !koinInject<AuthRepository>().accessTokenSync().isNullOrBlank()
         Surface(
